Fix abspos-in-block-in-inline-in-relpos-inline
This patch changes the test to avoid using
`Element.offsetWidth` to block-in-inline boxes.

`Element.offsetWidth` for block-in-inline is not interoperable.
Blink/WebKit ignores continuations, while Gecko takes them
into account. Also its spec[1] is not clear about how to
handle rects whose width or height is zero.

`getBounbingClientRects().width` is a possible replacement.
All 3 impls takes continuations into acccount, and its spec[2]
is clear about when width or height is zero, but WebKit has an
issue when width or height is zero.

This patch uses the size of inline-block instead as the
expected value to focus on what it wants to test.

I will add a separate test for `getBounbingClientRects()`.

[1] https://www.w3.org/TR/cssom-view-1/#dom-htmlelement-offsetwidth
[2] https://www.w3.org/TR/cssom-view-1/#dom-range-getboundingclientrect

Bug: 716930
Change-Id: I910a707f749513b231f6e41a9220f3b2a99be5be